# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for the Book Reader platform: public reading endpoints\n(books, chapters, verses, search, favorites, bookmarks, notes)\nand role-gated admin endpoints (books CRUD, Excel/CSV import,\nreference data, users, audit log).",
        "title": "Book Reader API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/categories/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "admin-categories"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "category id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.CategoryResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "admin-categories"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "category id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/admin/import/upload": {
            "post": {
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "admin-import"
                ],
                "summary": "Upload an Excel/CSV file and get a validation preview",
                "parameters": [
                    {
                        "type": "file",
                        "description": "xlsx, xls, or csv (Book, Chapter, Verse, text_en, text_id, title_en, title_id)",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ImportPreviewResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/import/{id}": {
            "get": {
                "tags": [
                    "admin-import"
                ],
                "summary": "Poll an import job's progress",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "import log id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ImportLogResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/import/{id}/commit": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "admin-import"
                ],
                "summary": "Start the actual import for a previously uploaded/validated file",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "import log id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Accepted",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.ImportLogResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/admin/import/{id}/stream": {
            "get": {
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "admin-import"
                ],
                "summary": "Server-Sent Events progress stream for an import job",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "import log id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {}
            }
        },
...
        "/admin/languages/{id}": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "admin-languages"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "language id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/response.Envelope"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/dto.LanguageResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "admin-languages"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "language id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
        "/admin/users/{id}": {
            "delete": {
                "tags": [
                    "admin-users"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/admin/users/{id}/role": {
            "put": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "admin-users"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "user id",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CategoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "name_en": {
                    "type": "string"
                },
                "name_id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                }
            }
        },
...
        "dto.ImportLogResponse": {
            "type": "object",
            "properties": {
                "books_created": {
                    "type": "integer"
                },
                "chapters_created": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "error_message": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "finished_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "mode": {
                    "type": "string"
                },
                "processed_rows": {
                    "type": "integer"
                },
                "sections_created": {
                    "type": "integer"
                },
                "started_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "total_rows": {
                    "type": "integer"
                },
                "verses_inserted": {
                    "type": "integer"
                },
                "verses_skipped": {
                    "type": "integer"
                },
                "verses_updated": {
                    "type": "integer"
                }
            }
        },
        "dto.ImportPreviewResponse": {
            "type": "object",
            "properties": {
                "distinct_books": {
                    "type": "integer"
                },
                "distinct_chapters": {
                    "type": "integer"
                },
                "import_log_id": {
                    "type": "integer"
                },
                "sample_rows": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ImportRowResponse"
                    }
                },
                "status": {
                    "type": "string"
                },
                "total_rows": {
                    "type": "integer"
                },
                "validation_errors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.ImportRowErrorResponse"
                    }
                }
            }
        },
        "dto.ImportRowErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "row_number": {
                    "type": "integer"
                }
            }
        },
        "dto.ImportRowResponse": {
            "type": "object",
            "properties": {
                "book": {
                    "type": "string"
                },
                "chapter": {
                    "type": "integer"
                },
                "row_number": {
                    "type": "integer"
                },
                "text_en": {
                    "type": "string"
                },
                "text_id": {
                    "type": "string"
                },
                "title_en": {
                    "type": "string"
                },
                "title_id": {
                    "type": "string"
                },
                "verse": {
                    "type": "integer"
                }
            }
        },
        "dto.LanguageResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "native_name": {
                    "type": "string"
                }
            }
        },
...
        "response.Envelope": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/response.ErrorBody"
                },
                "meta": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
        "response.ErrorBody": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "fields": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "CookieAuth": {
            "description": "JWT access token, set as an httpOnly cookie by /auth/login.",
            "type": "apiKey",
            "name": "access_token",
            "in": "cookie"
        }
    }
}